Todos los test pasan.

+++ Projects/P2/xv6-riscv/kernel/proc.h
@@ -93,8 +93,8 @@ struct proc {
   int xstate;                  // Exit status to be returned to parent's wait
   int pid;                     // Process ID
   int priority;                // priority 
-  uint64 hticks;               // clock ticks at high priority
-  uint64 lticks;               // clock ticks at low priority 
+  int hticks;               // clock ticks at high priority
+  int lticks;               // clock ticks at low priority 
 
 
   // wait_lock must be held when using this:
